switch to frame_benchmarking::v2
