+++ frame/support/test/tests/benchmark_ui/bad_param_name_too_long.rs
@@ -1,9 +1,9 @@
-use frame_benchmarking::*;
+use frame_benchmarking::v2::*;
 #[allow(unused_imports)]
 use frame_support_test::Config;
 
 #[benchmarks]
-mod benches {
+mod benchmarks {
 	#[benchmark]
 	fn bench(xx: Linear<1, 2>) {
 		#[block]
