+++ frame/support/test/tests/benchmark_ui/dup_extrinsic_call.rs
@@ -1,4 +1,4 @@
-use frame_benchmarking::*;
+use frame_benchmarking::v2::*;
 #[allow(unused_imports)]
 use frame_support_test::Config;
